refactor(server): split main.rs into focused modules
main.rs was 942 lines and contained CLI parsing, config setup,
foreground run, detached run, stop, status, interactive prompts,
background tasks (heartbeat & cleanup), and signal handling.

Split into focused modules:
- main.rs: CLI definition + entry point only (~44 lines)
- cli.rs: config preparation + interactive prompts
- cmd/run.rs: foreground server run + graceful shutdown
- cmd/detached.rs: detached background run
- cmd/stop.rs: stop server
- cmd/status.rs: server status query
- bg_tasks.rs: heartbeat checker + task cleanup

Fixes #54
