+++ agent-core/src/cmd/mod.rs
@@ -1,0 +1,6 @@
+pub mod run;
+pub mod detached;
+pub mod stop;
+pub mod status;
+pub mod init;
+pub mod register;
