+++ agent-core/src/cmd/register.rs
@@ -1,0 +1,108 @@
+use std::path::PathBuf;
+use agent_core::{client::ApiClient, config::Config, main_support::*};
+use tracing::{error, info};
+
+pub async fn register(
+    config_path: PathBuf,
+    token: Option<String>,
+    name: Option<String>,
+) -> anyhow::Result<()> {
+    info!("========================================");
+    info!("开始注册流程...");
+    info!("========================================");
+
+    // 加载配置
+    info!("步骤 1: 加载配置...");
+    let mut config = Config::load_from_path(&config_path).await?;
+    info!("配置加载成功: base_url={}", config.server.base_url);
+
+    // 检查是否已注册
+    info!("步骤 2: 检查是否已注册...");
+    if config.agent.has_credentials() {
+        println!(
+            "Agent 已注册: {}",
+            config.agent.service_id.as_ref().unwrap()
+        );
+        println!("如需重新注册，请删除配置文件后重试");
+        return Ok(());
+    }
+    info!("未注册，继续...");
+
+    // 设置名称
+    if let Some(name) = name {
+        config.agent.name = Some(name);
+    } else if config
+        .agent
+        .name
+        .as_deref()
+        .is_none_or(|value| value.trim().is_empty())
+    {
+        config.agent.name = Some("agent-core".to_string());
+    }
+
+    let token = match token {
+        Some(token) => token,
+        None if is_interactive_terminal() => prompt_registration_token()?,
+        None => anyhow::bail!("缺少注册令牌，请使用 agent-core register --token <TOKEN>"),
+    };
+
+    // 创建客户端
+    info!("步骤 3: 创建 API 客户端...");
+    let mut client = ApiClient::new(&config.server);
+    info!("API 客户端创建成功");
+
+    // 调用注册
+    info!("步骤 4: 调用 client.register()...");
+    match client.register(&token, config.executor.capacity).await {
+        Ok(response) => {
+            info!("注册成功: service_id={}", response.service_id);
+
+            // 保存配置
+            config.agent.api_key = Some(response.api_key);
+            config.agent.service_id = Some(response.service_id);
+            config.save_to_path(&config_path).await?;
+
+            println!("注册成功！");
+            println!("Service ID: {}", config.agent.service_id.as_ref().unwrap());
+            println!(
+                "API Key: {}...",
+                &config.agent.api_key.as_ref().unwrap()[..20]
+            );
+            println!("配置已保存");
+        }
+        Err(e) => {
+            error!("注册失败: {}", e);
+            eprintln!("注册失败！");
+            eprintln!("==============");
+            eprintln!("{}", e);
+            eprintln!();
+            eprintln!("常见问题:");
+            eprintln!("1. Server 地址配置错误: 检查 config.toml 中的 server.base_url");
+            eprintln!("2. 注册令牌过期: 联系管理员获取新的 registration_token");
+            eprintln!(
+                "3. Server 未运行: 确保 Server 在 {} 上监听",
+                config.server.base_url
+            );
+            std::process::exit(1);
+        }
+    }
+
+    Ok(())
+}
+
+pub fn prompt_registration_token() -> anyhow::Result<String> {
+    println!();
+    println!("--- Registration Token ---");
+    println!("Purpose: one-time token generated by the server when creating a service.");
+    println!("It is used only for first registration and is not saved by this prompt.");
+    println!();
+
+    loop {
+        let token = prompt_raw("Registration token", None)?;
+        let trimmed = token.trim().to_string();
+        if trimmed.starts_with("rt_") {
+            return Ok(trimmed);
+        }
+        eprintln!("Registration token should start with rt_.");
+    }
+}
