+++ server/src/cmd/mod.rs
@@ -1,0 +1,4 @@
+pub mod run;
+pub mod detached;
+pub mod stop;
+pub mod status;
